+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Remove button long press feature
 - Remove `pixcil_windows` crate to reduce maintenance costs
 
 ### Changed